+++ apps/client/src/app/pages/portfolio/activities/import-activities-dialog/import-activities-dialog.component.ts
@@ -154,9 +154,10 @@ export class GfImportActivitiesDialogComponent implements OnDestroy {
         })
         .pipe(takeUntil(this.unsubscribeSubject))
         .subscribe(({ holdings }) => {
-          this.holdings = sortBy(holdings, ({ name }) => {
-            return name.toLowerCase();
+          this.holdings = sortBy(holdings, ({ assetProfile }) => {
+            return assetProfile.name.toLowerCase();
           });
+
           this.assetProfileForm.get('assetProfileIdentifier').enable();
 
           this.isLoading = false;