+++ libs/ui/src/lib/assistant/assistant.component.ts
@@ -480,11 +480,14 @@ export class GfAssistantComponent implements OnChanges, OnDestroy, OnInit {
       .pipe(takeUntil(this.unsubscribeSubject))
       .subscribe(({ holdings }) => {
         this.holdings = holdings
-          .filter(({ assetSubClass }) => {
-            return assetSubClass && !['CASH'].includes(assetSubClass);
+          .filter(({ assetProfile }) => {
+            return (
+              assetProfile.assetSubClass &&
+              !['CASH'].includes(assetProfile.assetSubClass)
+            );
           })
           .sort((a, b) => {
-            return a.name?.localeCompare(b.name);
+            return a.assetProfile.name?.localeCompare(b.assetProfile.name);
           });
 
         this.setPortfolioFilterFormValues();
@@ -506,11 +509,11 @@ export class GfAssistantComponent implements OnChanges, OnDestroy, OnInit {
         type: 'ASSET_CLASS'
       },
       {
-        id: filterValue?.holding?.dataSource ?? '',
+        id: filterValue?.holding?.assetProfile?.dataSource ?? '',
         type: 'DATA_SOURCE'
       },
       {
-        id: filterValue?.holding?.symbol ?? '',
+        id: filterValue?.holding?.assetProfile?.symbol ?? '',
         type: 'SYMBOL'
       },
       {
@@ -697,18 +700,16 @@ export class GfAssistantComponent implements OnChanges, OnDestroy, OnInit {
           return EMPTY;
         }),
         map(({ holdings }) => {
-          return holdings.map(
-            ({ assetSubClass, currency, dataSource, name, symbol }) => {
-              return {
-                currency,
-                dataSource,
-                name,
-                symbol,
-                assetSubClassString: translate(assetSubClass ?? ''),
-                mode: SearchMode.HOLDING as const
-              };
-            }
-          );
+          return holdings.map(({ assetProfile }) => {
+            return {
+              assetSubClassString: translate(assetProfile.assetSubClass ?? ''),
+              currency: assetProfile.currency,
+              dataSource: assetProfile.dataSource,
+              mode: SearchMode.HOLDING as const,
+              name: assetProfile.name,
+              symbol: assetProfile.symbol
+            };
+          });
         }),
         takeUntil(this.unsubscribeSubject)
       );
@@ -752,12 +753,13 @@ export class GfAssistantComponent implements OnChanges, OnDestroy, OnInit {
       'filters.dataSource'
     ] as DataSource;
     const symbol = this.user?.settings?.['filters.symbol'];
-    const selectedHolding = this.holdings.find((holding) => {
+
+    const selectedHolding = this.holdings.find(({ assetProfile }) => {
       return (
         !!(dataSource && symbol) &&
         getAssetProfileIdentifier({
-          dataSource: holding.dataSource,
-          symbol: holding.symbol
+          dataSource: assetProfile.dataSource,
+          symbol: assetProfile.symbol
         }) === getAssetProfileIdentifier({ dataSource, symbol })
       );
     });